# LibJS: New bytecode format, remove JIT compiler