--- a/Meta/gn/secondary/Userland/Libraries/LibJS/BUILD.gn
+++ b/Meta/gn/secondary/Userland/Libraries/LibJS/BUILD.gn
@@ -48,8 +48,6 @@ shared_library("LibJS") {
     "Heap/Heap.cpp",
     "Heap/HeapBlock.cpp",
     "Heap/MarkedVector.cpp",
-    "JIT/Compiler.cpp",
-    "JIT/NativeExecutable.cpp",
     "Lexer.cpp",
     "MarkupGenerator.cpp",
     "Module.cpp",