--- a/Userland/Libraries/LibJS/Forward.h
+++ b/Userland/Libraries/LibJS/Forward.h
@@ -312,6 +312,7 @@ class Executable;
 class Generator;
 class Instruction;
 class Interpreter;
+class Operand;
 class RegexTable;
 class Register;
 }