… .github/workflows/libjs-test262.yml
… Userland/Libraries/LibJS/AST.h
… Userland/Libraries/LibJS/Bytecode/ASTCodegen.cpp
… Userland/Libraries/LibJS/Bytecode/Interpreter.cpp
… Userland/Libraries/LibJS/Bytecode/Interpreter.h
… Userland/Libraries/LibJS/Bytecode/Op.h
… Userland/Libraries/LibJS/Bytecode/Operand.h
… Userland/Libraries/LibJS/CMakeLists.txt
… Userland/Libraries/LibJS/JIT/Compiler.cpp
… Userland/Libraries/LibJS/JIT/Compiler.h
… Userland/Libraries/LibJS/JIT/NativeExecutable.cpp
… Userland/Libraries/LibJS/JIT/NativeExecutable.h
… Userland/Libraries/LibJS/Runtime/AbstractOperations.cpp
… Userland/Libraries/LibJS/Runtime/ECMAScriptFunctionObject.cpp
… Userland/Libraries/LibJS/Runtime/ShadowRealm.cpp
… Userland/Libraries/LibJS/Runtime/VM.cpp
… Userland/Libraries/LibJS/Runtime/VM.h
… Userland/Libraries/LibJS/SourceTextModule.cpp