@skyrising 2024-02-19
Love the new bytecode format, but as someone who poured a bit of work into the JIT it's sad to see it go.

While a 1.25x speedup is nice, the JIT did provide a ~2.4x speedup (5.4x Kraken, 2.3x Octane), so I can't help but wonder what these the combination of new bytecode and JIT could achieve.

Wouldn't it be nicer to just disable the JIT for now, allowing somebody to fix it up, instead of outright removing it? (We could still remove it later if nobody wants to put in the work).

Obviously removing it is way easier than keeping it working at the moment, but that also makes it harder for anyone wanting to bring it back in the future.