@awesomekling 2024-02-19
> Wouldn't it be nicer to only disable the JIT for now, allowing somebody to fix it up, instead of outright removing it? (We could still remove it later if nobody wants to put in the work).

The JIT is still there in git history for anyone who wants to do something with it. :)