@awesomekling 2024-02-19
> The last commit is a lot to chew but seems good to me! At least I don't seem anything glaringly out of sorts.
> 
> A couple things that stuck out to me: ...

Took care of all your comments, I think!